+++ libraries/js/src/index.ts
@@ -1,4 +1,5 @@
 export * from './request.js';
 export * from './response.js';
 export * from './start.js';
+export * from './constants.js';
 export * from './types/index.js';
